fix(#281): add serviceDir to Serverless config
serviceDir and servicePath both are optional so we have to check.
In index.ts, I'm setting both, because according to comments throughout Serverless code, they should actually
be linked. Due to (I think) a bug, they're not. But there may be code that uses either/or, so they should both be set.
Added setter and getter for this, to put that logic in one place.

This doesn't fix #281 yet, but it's a necessary part.